+++ Fuji/Private/Source/FileSystem/MFFileSystemMemory_Internal.h
@@ -6,7 +6,7 @@
 struct MFFileMemoryData
 {
 	void *pMemoryPointer;
-	uint32 allocated;
+	size_t allocated;
 	bool ownsMemory;	// this calls Heap_Free() on the memory pointer when the file is closed.
 };
