+++ Fuji/Private/Source/MFIni.cpp
@@ -155,7 +155,7 @@ MFIni *MFIni::Create(const char *pFilename)
 {
 	MFCALLSTACK;
 
-	uint32 memSize;
+	size_t memSize;
 
 	// load text file
 #if !defined(_FUJI_UTIL)
@@ -187,7 +187,7 @@ MFIni *MFIni::Create(const char *pFilename)
 	// allocate temporary buffer for strings & lines
 	pMFIni->linesAllocated = 2048;
 	pMFIni->stringsAllocated = 2048;
-	pMFIni->stringCacheSize = memSize+2+(7*4); // 7*4 = memory for a couple of "section" strings..
+	pMFIni->stringCacheSize = (int)memSize+2+(7*4); // 7*4 = memory for a couple of "section" strings..
 	pMFIni->pLines = (MFIniLine*)MFHeap_Alloc(sizeof(MFIniLine)*pMFIni->linesAllocated);
 	pMFIni->pStrings = (const char**)MFHeap_Alloc(sizeof(const char*)*pMFIni->stringsAllocated);
 	pMFIni->pCache = MFStringCache_Create(pMFIni->stringCacheSize);
