+++ Haku/Source/UI/HKWidgetStyle.cpp
@@ -21,7 +21,7 @@ void HKWidgetStyle::Deinit()
 bool HKWidgetStyle::LoadStylesFromXML(const char *pFilename)
 {
 	// attempt to load the xml document
-	uint32 len;
+	size_t len;
 	char *pFile = MFFileSystem_Load(pFilename, &len, true);
 
 	if(!pFile)
